refactor(pipeline): address CodeRabbit review findings
- connection_manager: broadcast over a snapshot of the connection set to
  avoid 'Set changed size during iteration' when a client disconnects
  concurrently during await send_json
- ingest: offload synchronous Redis/RQ submit work to a worker thread via
  run_in_threadpool so blocking I/O stays off the event loop
- bridge: type _relay's opaque payload as object (ANN401) instead of Any
- tests: move integration-marked lifespan test to tests/integration
